Split the python code into a stage class and a Sangaboard class
