+++ openflexure_stage/stage.py
@@ -10,14 +10,11 @@ It is (c) Richard Bowman 2017 and released under GNU GPL v3
 """
 from __future__ import print_function, division
 import time
-from .extensible_serial_instrument import ExtensibleSerialInstrument, OptionalModule, QueriedProperty, EIGHTBITS, PARITY_NONE, STOPBITS_ONE
+from .sangaboard import Sangaboard
 import numpy as np
-import sys
-import re
-import warnings
 
-class OpenFlexureStage(ExtensibleSerialInstrument):
-    """Class managing serial communications with an Openflexure Motor Controller
+class OpenFlexureStage():
+    """Class managing serial communications with the motors for an Openflexure stage
 
     The `OpenFlexureStage` class handles setting up communications with the stage,
     wraps the various serial commands in Python methods, and provides iterators and
@@ -38,82 +35,89 @@ class OpenFlexureStage(ExtensibleSerialInstrument):
     even if an error occurs.  Otherwise, be sure to call the
     :meth:`~.ExtensibleSerialInstrument.close()` method to release the serial port.
     """
-    port_settings = {'baudrate':115200, 'bytesize':EIGHTBITS, 'parity':PARITY_NONE, 'stopbits':STOPBITS_ONE}
-    """These are the settings for the stage's serial port, and can usually be left as default."""
-    # position, step time and ramp time are get/set using simple serial
-    # commands.
-    position = QueriedProperty(get_cmd="p?", response_string=r"%d %d %d",
-            doc="Get the position of the stage as a tuple of 3 integers.")
-    step_time = QueriedProperty(get_cmd="dt?", set_cmd="dt %d", response_string="minimum step delay %d",
-            doc="Get or set the minimum time between steps of the motors in microseconds.\n\n"
-                "The step time is ``1000000/max speed`` in steps/second.  It is saved to EEPROM on "
-                "the Arduino, so it will be persistent even if the motor controller is turned off.")
-    ramp_time = QueriedProperty(get_cmd="ramp_time?", set_cmd="ramp_time %d", response_string="ramp time %d",
-            doc="Get or set the acceleration time in microseconds.\n\n"
-                "The stage will accelerate/decelerate between stationary and maximum speed over `ramp_time` "
-                "microseconds.  Zero means the stage runs at full speed initially, with no accleration "
-                "control.  Small moves may last less than `2*ramp_time`, in which case the acceleration "
-                "will be the same, but the stage will never reach full speed.  It is saved to EEPROM on "
-                "the Arduino, so it will be persistent even if the motor controller is turned off.")
-    axis_names = ('x', 'y', 'z')
-    """The names of the stage's axes.  NB this also defines the number of axes."""
+    
     board = None
-    """Once initialised, `board` is a string that identifies the firmware version."""
+    """Once initialised, `board` is is the motor controller board"""
+    boardtype = None
     supported_light_sensors = ["TSL2591","ADS1115"]
     """This is a list of the supported light sensor module types."""
+    
+    SANGABOARD = 0
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, port=None,board='Sangaboard', **kwargs):
         """Create a stage object.
 
-        Arguments are passed to the constructor of
-        :class:`openflexure_stage.extensible_serial_instrument.ExtensibleSerialInstrument`,
+        port: The port on which is used to communicat to the motor board
+        
+        board: Sets the motor board used in the openflexure stage. Currently Sangaboard is
+        the only board supported.
+        
+        Other arguments are passed to the constructor of the class for the selected board
         most likely the only one necessary is `port` which should be set to the serial port
         you will use to communicate with the motor controller.  That's the first argument so
         it doesn't need to be named.
         """
-        super(OpenFlexureStage, self).__init__(*args, **kwargs)
-        try:
-            # Request version from the board
-            self.board = self.query("version").rstrip()
-            # The slightly complicated regexp below will match the version string,
-            # and store the version number in the "groups" of the regexp.  The version
-            # number should be in the format 1.2 and the groups will be "1.2", "1", "2"
-            # (for any number of elements).
-            match = re.match(r"OpenFlexure Motor Board v(([\d]+)(?:\.([\d]+))+)", self.board)
-            version = [int(g) for g in match.groups()[1:]]
-            assert match, "Version string \"{}\" not recognised.".format(self.board)
-            self.firmware_version = match.group(1)
-            assert version[0] == 0, "This version of the Python module requires firmware v0.4"
-            assert version[1] == 4, "This version of the Python module requires firmware v0.4"
-
-            #Bit messy: Defining all valid modules as not available, then overwriting with available information if available.
-            self.light_sensor = LightSensor(False)
-
-            for module in self.list_modules():
-                module_type=module.split(':')[0].strip()
-                module_model=module.split(':')[1].strip()
-                if module_type.startswith('Light Sensor'):
-                    if module_model in self.supported_light_sensors:
-                        self.light_sensor = LightSensor(True,parent=self,model=module_model)
-                    else:
-                        warnings.warn("Light sensor model \"%s\" not recognised."%(module_model),RuntimeWarning)
-                elif module_type.startswith('Endstops'):
-                    self.endstops = Endstops(True, parent=self, model=module_model)
-                else:
-                    warnings.warn("Module type \"{}\" not recognised.".format(module_type),RuntimeWarning)
-            self.test_mode=False
+        
+        if board.startswith('Sangaboard'):
+            self.board = Sangaboard(port,**kwargs)
+            self.boardtype = self.SANGABOARD
+        else:
+            assert False, "Board type not recognised"
+        
         except Exception as e:
             # If an error occurred while setting up (e.g. because the board isn't connected or something)
             # make sure we close the serial port cleanly (otherwise it hangs open).
             self.close()
-            e.args += ("You may need to update the firmware running on the Arduino.  See " \
+            e.args += ("You may need to update the firmware running on the motorboard.  See " \
                        "https://openflexure-stage.readthedocs.io/en/latest/firmware.html",)
             raise e
+    
+    
+    def open(self, port=None, quiet=True):
+        """Open communications with the serial port.
+        
+        If no port is specified, it will attempt to autodetect.  If quiet=True
+        then we don't warn when ports are opened multiple times.
+        """
+        self.board.open(port,quiet)
+    
+    def close(self):
+        self.board.close()
 
+    def __del__(self):
+        """Close the port when the object is deleted
+        
+        NB if the object is created in a with statement, this will cause
+        the port to be closed at the end of the with block."""
+        self.close()
+
+    def __enter__(self):
+        """When we use this in a with statement, remember where we started."""
+        self._position_on_enter = self.position
+        return self
+
+    def __exit__(self, type, value, traceback):
+        """The end of the with statement.  Reset position if it went wrong.
+        NB the instrument is closed when the object is deleted, so we don't
+        need to worry about that here.
+        """
+        if type is not None:
+            print("An exception occurred inside a with block, resetting " +
+                  "position to its value at the start of the with block")
+            try:
+                time.sleep(0.5)
+                self.move_abs(self._position_on_enter)
+            except Exception as e:
+                print("A further exception occurred when resetting position: {}".format(e))
+            print("Move completed, raising exception...")
+            raise value #propagate the exception
+    
+    
+    
     @property
     def n_axes(self):
         """The number of axes this stage has."""
-        return len(self.axis_names)
+        return len(self.board.axis_names)
 
     _backlash = None
     @property
@@ -155,7 +159,7 @@ class OpenFlexureStage(ExtensibleSerialInstrument):
         backlash: (default: True) whether to correct for backlash.
         """
         if not backlash or self.backlash is None:
-            return self._move_rel_nobacklash(displacement, axis=axis)
+            return self.board.move_rel(displacement, axis=axis)
 
         if axis is not None:
             # backlash correction is easier if we're always in 3D
@@ -177,38 +181,20 @@ class OpenFlexureStage(ExtensibleSerialInstrument):
         initial_move -= np.where(self.backlash*displacement < 0,
                                  self.backlash,
                                  np.zeros(self.n_axes, dtype=self.backlash.dtype))
-        self._move_rel_nobacklash(initial_move)
+        self.board.move_rel(initial_move)
         if np.any(displacement - initial_move != 0):
             # If backlash correction has kicked in and made us overshoot, move
             # to the correct end position (i.e. the move we were asked to make)
-            self._move_rel_nobacklash(displacement - initial_move)
-
-    def _move_rel_nobacklash(self, displacement, axis=None):
-        """Just make a move - no messing about with backlash correction!
-
-        Arguments are as for move_rel, but backlash is False
-        """
-        if axis is not None:
-            assert axis in self.axis_names, "axis must be one of {}".format(self.axis_names)
-            self.query("mr{} {}".format(axis, int(displacement)))
-        else:
-            #TODO: assert displacement is 3 integers
-            self.query("mr {} {} {}".format(*list(displacement)))
+            self.board.move_rel(displacement - initial_move)
 
     def release_motors(self):
         """De-energise the stepper motor coils"""
-        self.query("release")
+        self.board.release_motors()
 
     def move_abs(self, final, **kwargs):
         """Make an absolute move to a position
-
-        NB the stage only accepts relative move commands, so this first
-        queries the stage for its position, then instructs it to make about
-        relative move.
         """
-        new_position = final
-        rel_mov = np.subtract(new_position, self.position)
-        return self.move_rel(rel_mov, **kwargs)
+        self.board.move_abs(final, **kwargs)
 
     def focus_rel(self, z):
         """Move the stage in the Z direction by z micro steps."""
@@ -253,43 +239,18 @@ class OpenFlexureStage(ExtensibleSerialInstrument):
         """
         return self.scan_linear([[0,0,z] for z in dz], **kwargs)
 
-    def __enter__(self):
-        """When we use this in a with statement, remember where we started."""
-        self._position_on_enter = self.position
-        return self
-
-    def __exit__(self, type, value, traceback):
-        """The end of the with statement.  Reset position if it went wrong.
-        NB the instrument is closed when the object is deleted, so we don't
-        need to worry about that here.
-        """
-        if type is not None:
-            print("An exception occurred inside a with block, resetting " +
-                  "position to its value at the start of the with block")
-            try:
-                time.sleep(0.5)
-                self.move_abs(self._position_on_enter)
-            except Exception as e:
-                print("A further exception occurred when resetting position: {}".format(e))
-            print("Move completed, raising exception...")
-            raise value #propagate the exception
-
     def query(self, message, *args, **kwargs):
-        """Send a message and read the response.  See ExtensibleSerialInstrument.query()"""
-        time.sleep(0.001) # This is to protect the stage from us talking too fast!
-        return ExtensibleSerialInstrument.query(self, message, *args, **kwargs)
+        self.board.query( message, *args, **kwargs)
 
     def list_modules(self):
         """Return a list of strings detailing optional modules.
 
         Each module will correspond to a string of the form ``Module Name: Model``
         """
-        modules =  self.query("list_modules",multiline=True,termination_line="--END--\r\n").split('\r\n')[:-2]
-        return [str(module) for module in modules]
+        return self.board.list_modules()
 
     def print_help(self):
-        """Print the stage's built-in help message."""
-        print(self.query("help",multiline=True,termination_line="--END--\r\n"))
+        self.board.print_help()
 
     @property
     def test_mode(self):
@@ -311,115 +272,4 @@ class OpenFlexureStage(ExtensibleSerialInstrument):
             self.endstops.test_mode=value
         if hasattr(self, 'light_sensor') and self.light_sensor:
             self.light_sensor.test_mode=value
-
-
-
-class LightSensor(OptionalModule):
-    """An optional module giving access to the light sensor.
-
-    If a light sensor is enabled in the motor controller's firmware, then
-    the :class:`openflexure_stage.OpenFlexureStage` will gain an optional
-    module which is an instance of this class.  It can be used to access
-    the light sensor (usually via the I2C bus).
-    """
-    valid_gains = None
-    _valid_gains_int = None
-    integration_time = QueriedProperty(get_cmd="light_sensor_integration_time?",
-                                       set_cmd="light_sensor_integration_time %d",
-                                       response_string="light sensor integration time %d ms",
-                                       doc="Get or set the integration time of the light sensor in milliseconds.")
-    intensity = QueriedProperty(get_cmd="light_sensor_intensity?", response_string="%d",
-                                doc="Read the current intensity measured by the light sensor (arbitrary units).")
-
-    def __init__(self,available,parent=None,model="Generic"):
-        super(LightSensor, self).__init__(available,parent=parent,module_type="LightSensor",model=model)
-        if available:
-            self.valid_gains = self.__get_gain_values()
-            self._valid_gains_int = [int(g) for g in self.valid_gains]
-
-    @property
-    def gain(self):
-        """"Get or set the current gain value of the light sensor.
-
-        Valid gain values are defined in the `valid_gains` property, and should be floating-point numbers."""
-        self.confirm_available()
-        gain = self._parent.query('light_sensor_gain?')
-        M = re.search('[0-9\.]+(?=x)',gain)
-        assert M is not None, "Cannot read gain string: \"{}\"".format(gain)
-        #gain is a float as non integer gains exist but are set with floor of value
-        return float(M.group())
-
-    @gain.setter
-    def gain(self,val):
-        self.confirm_available()
-        assert int(val) in self._valid_gains_int, "Gain {} not valid must be one of: {}".format(val,self.valid_gains)
-        gain = self._parent.query('light_sensor_gain %d'%(int(val)))
-        M = re.search('[0-9\.]+(?=x)',gain)
-        assert M is not None, "Cannot read gain string: \"{}\"".format(gain)
-        #gain is a float as non integer gains exist but are set with floor of value
-        assert int(val) == int(float(M.group())), 'Gain of {} set, \"{}\" returned'.format(val,gain)
-
-    def __get_gain_values(self):
-        """Read the allowable values for the light sensor's gain.
-
-        This function will attempt to return a list of floating-point numbers which may
-        be used as values of the `gain` property.  If the stage returns non-floating-point
-        values, the list will be of strings.
-        """
-        self.confirm_available()
-        gains = self._parent.query('light_sensor_gain_values?')
-        try:
-            M = re.findall('[0-9\.]+(?=x)',gains)
-            return [float(gain) for gain in M]
-        except:
-            # Fall back to strings if we don't get floats (unlikely)
-            gain_strings = gains[20:].split(", ")
-            return gain_strings
-
-class Endstops(OptionalModule):
-    """An optional module for use with endstops.
-
-    If endstops are installed in the firmware the :class:`openflexure_stage.OpenFlexureStage`
-    will gain an optional module which is an instance of this class.  It can be used to retrieve
-    the type, state of the endstops, read and write maximum positions, and home.
-    """
-
-    installed=[]
-    test_mode=False
-    """List of installed endstop types (min, max, soft)"""
-
-    def __init__(self,available,parent=None,model="min"):
-        super(Endstops, self).__init__(available,parent=parent,model="Endstops")
-        self.installed=model.split(' ')
-
-    status = QueriedProperty(get_cmd="endstops?", response_string=r"%d %d %d",
-                            doc="Get endstops status as {-1,0,1} for {min,no,max} endstop triggered for each axis")
-    maxima = QueriedProperty(get_cmd="max_p?", set_cmd="max %d %d %d", response_string="%d %d %d",
-                            doc="Vector of maximum positions, homing to max endstops will measure this, "+
-                                "can be set to a known value for use with max only and min+soft endstops")
-
-    def home(self, direction="min", axes=['x','y','z']):
-        """ Home given/all axes in the given direction (min/max/both)
-
-            :param direction: one of {min,max,both}
-            :param axes: list of axes e.g. ['x','y']
-        """
-        ax=0
-        if 'x' in axes:
-            ax+=1
-        if 'y' in axes:
-            ax+=2
-        if 'z' in axes:
-            ax+=3
-
-        if direction == "min" or direction == "both":
-            if self.test_mode:
-                return self._parent.query('home_min {}'.format(ax),multiline=True,termination_line="done.\r\n")
-            else:
-                self._parent.query('home_min {}'.format(ax))
-        if direction == "max" or direction == "both":
-            if self.test_mode:
-                return self._parent.query('home_max {}'.format(ax),multiline=True,termination_line="done.\r\n")
-            else:
-                self._parent.query('home_max {}'.format(ax))
